% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnetPlot_from_enrichmentResult.R
\name{cnetplot.enrichResult}
\alias{cnetplot.enrichResult}
\title{cnetplot.enrichResult}
\usage{
cnetplot.enrichResult(x, showCategory = 5, categorySize = "pvalue",
  foldChange = NULL, fixed = TRUE, ...)
}
\arguments{
\item{x}{The dataframe of enrichment analysis result}

\item{showCategory}{The number of enriched term to show}

\item{categorySize}{column name of x that used to scale the GO term. defult is "pvalue" scale the categorySize by enrichment pvalue}
}
