% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeakFisher.R
\name{.fisher_exact_test}
\alias{.fisher_exact_test}
\title{Fisher's exact test}
\usage{
.fisher_exact_test(IP, input, IP_overall, input_overall, pseudo_count = 1)
}
\arguments{
\item{counts.m}{matrix with the first column as IP counts, and second column as input counts}
}
\value{
data frame with p-values and odds ratio
}
\description{
Fisher's exact test
}
