% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPeaks.R
\name{filterPeak}
\alias{filterPeak}
\title{filterPeak}
\usage{
filterPeak(readsOut, minINPUT = 15, minIP = 30)
}
\arguments{
\item{readsOut}{The data list contain peak calling result}

\item{minINPUT}{Define the minimal number of INPUT reads (median across samples).}

\item{minIP}{Define the minimal number of IP reads (median across samples)}
}
\value{
readsOut with updated jointPeak_id_pairs that have been filtered by read count.
}
