% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTPM.R
\name{getTPM}
\alias{getTPM}
\title{getTPM}
\usage{
getTPM(x, meanFragmentLength = 150, normalize = T)
}
\arguments{
\item{x}{The m6Amonster object}

\item{meanFragmentLength}{The mean length of RNA fragment (insert of RNA library). Default is 150bp.}

\item{normalize}{Logical indicating whether normalized TPM or raw TPM should be returned.}
}
