% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuitarPlot2.R
\name{GuitarPlot}
\alias{GuitarPlot}
\title{GuitarPlot}
\usage{
GuitarPlot(gfeatures, GuitarCoordsFromTxDb = NA, txdb = NA, genome = NA,
  noBins = 10, saveToPDFprefix = NA, returnCount = FALSE,
  includeNeighborDNA = FALSE, maximalFeatureAmbiguity = 5,
  rescaleComponent = TRUE, fill = FALSE, adjust = 1)
}
