% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointPeakCount.R
\name{JointPeakCount}
\alias{JointPeakCount}
\title{JointPeakCount}
\usage{
JointPeakCount(readsOut, joint_threshold = 2)
}
\arguments{
\item{readsOut}{The data list as output of callPeakFisher()}

\item{joint_threshold}{Require a bin to be called as a peak in joint_threshold number of samples to be a joint peak.}
}
