% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpression.R
\name{plotExpression}
\alias{plotExpression}
\title{plotExpression}
\usage{
plotExpression(RDM, geneName, logCount = FLASE)
}
\arguments{
\item{RDM}{The RDM object}

\item{geneName}{The name of genes to be ploted.}

\item{logCount}{where to plot count at log scale}
}
