% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGene.R
\name{plotGeneMonster}
\alias{plotGeneMonster}
\title{plotGeneMonster}
\usage{
plotGeneMonster(readsOut, geneName, libraryType = "opposite",
  center = "mean", ZoomIn = NULL)
}
\arguments{
\item{readsOut}{The data list from countReads and other analysis.}
}
