% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.DESeq2.R
\name{plotPCA.DESeq2}
\alias{plotPCA.DESeq2}
\title{plotPCA.DESeq2}
\usage{
plotPCA.DESeq2(data, group = NULL, returnPC = FALSE)
}
\arguments{
\item{data}{The data matrix}

\item{group}{The label of each column of the data.}
}
\description{
Takes a matrix and regularize transform the data using DESeq2::rlog function. Then plot PCA using the DESeq2.
}
