% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportConsistentPeak.R
\name{reportConsistentPeak}
\alias{reportConsistentPeak}
\title{reportConsistentPeak}
\usage{
reportConsistentPeak(readsOut, samplenames, joint_threshold = NULL,
  threads = 1)
}
\arguments{
\item{readsOut}{The data list contain peak calling result}

\item{samplenames}{The samplenames to be reported for consistent peaks}

\item{joint_threshold}{Define the number of sample required to have consistent peak in a locus to call consistent peak in a group.
This option is useful only when joint poeak has not been reported by jointPeakCount().}
}
\value{
merged.report The joint peak of merged bins.
}
