% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportJointPeak.R
\name{reportJointPeak}
\alias{reportJointPeak}
\title{reportJointPeak}
\usage{
reportJointPeak(readsOut, joint_threshold, threads = 1)
}
\arguments{
\item{readsOut}{The data list contain peak calling result}

\item{joint_threshold}{Define the number of sample required to have consistent peak in a locus to call joint peak.
This option is useful only when joint poeak has not been reported by jointPeakCount().}
}
\value{
merged.report The joint peak of merged bins.
}
