% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDistribution.R
\name{peakDistribution}
\alias{peakDistribution}
\title{gtfToGeneModel}
\usage{
peakDistribution(peak, gtf, saveName = NA)
}
\arguments{
\item{gtf}{gtf file to build gene model}

\item{saveName}{the file name to save ditribution plot}

\item{bedFile}{peak list as bed format}
}
\description{
to remove ambiguous gene model and return gene model as genomic ranges object
}
